refactor(server): move agent routes into package
